Add support for multiple keymap bindings to qtile.
If you do not want to use this feature, no configuration changes are
necessary.

Add libqtile.config.KeyMap which is a list with some additional
attributes describing the keymap, and some convenience methods
for binding keys.

Add the libqtile.widget.CurrentKeyMap widget, which displays
the name of the current keymap.